requirements.txt
matplotlib==3.8.1
numpy==1.26.1
pandas==2.1.2
plotly==5.18.0
pytest==7.4.3
PyYAML==6.0.1
scikit-learn==1.3.2
scrapy==2.11.0
seaborn==0.13.0
spacy==3.7.2
umap-learn==0.5.4
beautifulsoup4==4.12.2
termcolor==2.3.0
python-box==7.1.1
geopandas==0.14.0
unicode==2.9
xgboost==2.0.1
shap==0.43.0
bayesian-optimization==1.4.3
optuna==3.4.0
httpx==0.25.1
aiohttp==3.8.6
nest-asyncio==1.5.8
mlflow==2.8.0